requirements.txt
# --- Core AI (100% local; no API keys) ---
openai-whisper      # Local Whisper model (runs on your machine, no OpenAI API)
ollama              # Interface for local Llama 3.2 via Ollama

# --- Media & Processing ---
yt-dlp              # Downloading audio from YouTube
pydub               # Cutting and manipulating audio
music-tag           # Adding metadata (ID3 tags) to audio files

# --- Document Generation ---
fpdf2               # Creating the printable Wisdom PDFs

# --- Utilities (optional) ---
# python-dotenv     # Uncomment if you add optional env-based config